Refactor access code to AccessUtil and rename accessPredicate -> accessFilter
